Improves the Auto Layout Margins example
